Replace Utils subscriber functions with Subscriber Registry
Moving the registry for a subscriber to a new module so that we can
inject some options easier when test.

Amos King @adkron <amos@binarynoggin.com>
